+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not released
 
 - Fix: table widget was not showing data due to case sensitive [#721](https://github.com/CartoDB/carto-react/pull/721)
+- Supporting for `client` parameter for Widgets API calls [#722](https://github.com/CartoDB/carto-react/pull/722)
 
 ## 2.1